requirements.txt
openai-whisper
sounddevice
numpy
PyQt6
